notes_api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
databases[sqlite]==0.9.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
